% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{adjust}
\alias{adjust}
\title{Transform an input variable to be uncorrelated with other variables.}
\usage{
adjust(x, in_var, adj_var, loops = NULL)
}
\arguments{
\item{x}{data frame, containing all the necessary input data.}

\item{in_var}{character vector, the name of the variable to be 
\code{adjust}ed.}

\item{adj_var}{character vector, the name of the variable(s) to be used in 
the \code{adjust}ment process.}

\item{loops}{numeric scalar, the number of times to run the pair of 
adjustment regressions.  The default is \code{NULL}, which just runs the 
equal-weighted regression a single time.}
}
\value{
a numeric vector equal in length to \code{in_var} which is 
  uncorrelated with all the variables in \code{adj_var}.
}
\description{
\code{adjust} makes \code{in_var} neutral (uncorrelated with) the variables 
in \code{adj_var}. Complexity arises because we should ensure that all the 
variables used have normal distributions. Also, we generally want to be 
neutral in two senses: first, uncorrelated with each \code{adj_var} and, 
second, no exposure to any \code{adj_var} when looking at decile/quintile 
spreads of the output variable. Of course, we would also like the output 
variable to be highly correlated with \code{in_var}. We achieve this by 
running both equal-weighted and sigmoid-weighted regressions in a loop.
}
