% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMENDS_Data.R
\docType{data}
\name{AMENDS_Data}
\alias{AMENDS_Data}
\title{Friendship and sharing network data from a rural Colombian population}
\format{
An object of class \code{"list"}.
}
\source{
\href{https://www.eva.mpg.de/ecology/fieldwork/afrocolombians}{EVA}
}
\usage{
data(AMENDS_Data)
}
\description{
Data on friendship nominations, sharing ties, relatedness, age, sex, wealth, and ethnicity.
}
\examples{
data(AMENDS_Data)
}
\references{
Ross, CT., and Pisor, A.
2024. Perceived inequality and variability in the expression of parochial altruism.
Evolutionary Human Sciences.
}
\keyword{datasets}
