% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Baboon_Data.R
\docType{data}
\name{Baboon_Data}
\alias{Baboon_Data}
\title{Data on Baboon grooming relationships}
\format{
An object of class \code{"list"}.
}
\source{
\href{http://dx.doi.org/10.1098/rspa.2019.0737}{RSPA}
}
\usage{
data(Baboon_Data)
}
\description{
Data on grooming paired with data on presenting. Exposure was calculated from total times each indivudal was seen doing the behaviors of interest.
}
\examples{
data(Baboon_Data)
}
\references{
Gelardi V, Godard J, Paleressompoulle D, Claidiere N, Barrat A. 2020.
Measuring social networks in primates: wearable sensors versus direct observations.
Proc. R. Soc. A 476: 20190737.
}
\keyword{datasets}
