% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bat_Data.R
\docType{data}
\name{Bat_Data}
\alias{Bat_Data}
\title{Blood sharing data in vampire bats}
\format{
An object of class \code{"list"}.
}
\source{
\href{http://dx.doi.org/10.1098/rspb.2012.2573}{RSPB}
}
\usage{
data(Bat_Data)
}
\description{
Data on blood sharing observations. Kinship data is based on SPAGeDi software estimates from the authors STR data.
}
\examples{
data(Bat_Data)
}
\references{
Carter GG, Wilkinson GS.
2013. Food sharing in vampire bats: reciprocal help predicts donations more than relatedness or harassment. 
Proc R Soc B 280: 20122573.
}
\keyword{datasets}
