% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_p.R
\name{bayesian_p}
\alias{bayesian_p}
\title{A function to calculate prob of sign error, or Bayesian p value}
\usage{
bayesian_p(x)
}
\arguments{
\item{x}{value}
}
\value{
bayesian_p(x)
}
\description{
A function to calculate prob of sign error, or Bayesian p value
}
