% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_set_to_dyadic_block_set.R
\name{block_set_to_dyadic_block_set}
\alias{block_set_to_dyadic_block_set}
\title{A data-parser function to translate a matrix of Block ID variables into a tensor of indicators}
\usage{
block_set_to_dyadic_block_set(x, priors = NULL)
}
\arguments{
\item{x}{A data frame of indicators.}

\item{priors}{A data frame of priors.}
}
\value{
block_set_to_dyadic_block_set(x)
}
\description{
This is an internal function.
}
