% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/build_multiplex_bindings_dr_longitudinal_numpyro.R
\name{build_multiplex_bindings_dr_longitudinal_numpyro}
\alias{build_multiplex_bindings_dr_longitudinal_numpyro}
\title{A function to add accesssory data to a STRAND data object for use in longitudinal analysis}
\usage{
build_multiplex_bindings_dr_longitudinal_numpyro(data, verbose = FALSE)
}
\arguments{
\item{data}{A STRAND data object.}

\item{verbose}{If TRUE, then print structure of the bindings.}
}
\value{
A STRAND data object with extra information about the structure of the dyadic reciprocity matrix.
}
\description{
This is an internal function.
}
