% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_multiplex_bindings_dr_multiplex.R
\name{build_multiplex_bindings_dr_multiplex}
\alias{build_multiplex_bindings_dr_multiplex}
\title{A function to add accesssory data to a STRAND data object for use in multiplex analysis}
\usage{
build_multiplex_bindings_dr_multiplex(data, verbose = FALSE)
}
\arguments{
\item{data}{A STRAND data object.}

\item{verbose}{If TRUE, then print structure of dyadic reciprocity matrix.}
}
\value{
A STRAND data object with extra information about the structure of the dyadic reciprocity matrix.
}
\description{
This is an internal function.
}
