% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_multiplex_bindings_sr_longitudinal.R
\name{build_multiplex_bindings_sr_longitudinal}
\alias{build_multiplex_bindings_sr_longitudinal}
\title{A function to add accesssory data to a STRAND data object for use in longitudinal analysis}
\usage{
build_multiplex_bindings_sr_longitudinal(data, verbose = FALSE)
}
\arguments{
\item{data}{A STRAND data object.}

\item{verbose}{If TRUE, then print structure of generalized reciprocity matrix.}
}
\value{
A STRAND data object with extra information about the structure of the generalized reciprocity matrix.
}
\description{
This is an internal function.
}
