% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Callithrix_Data.R
\docType{data}
\name{Callithrix_Data}
\alias{Callithrix_Data}
\title{Data on Callithrix agression}
\format{
An object of class \code{"list"}.
}
\source{
\href{https://doi.org/10.5061/dryad.7pvmcvdtn}{Dryad}
}
\usage{
data(Callithrix_Data)
}
\description{
Foraging networks and social tolerance in a cooperatively breeding primate (Callithrix jacchus)
}
\examples{
data(Callithrix_Data)
}
\references{
De la Fuente, María Fernanda et al. (2021), Foraging networks and social tolerance in a cooperatively breeding primate (Callithrix jacchus), Dryad, Dataset.
}
\keyword{datasets}
