% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{A function to center data}
\usage{
center(input)
}
\arguments{
\item{input}{A vector of data to center.}
}
\value{
A vector of centered data.
}
\description{
This is a simple helper function to center data, as it makes parameter interpretation easier.
}
\examples{
\dontrun{
age_centered = center(input=age)
}

}
