% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CES.R
\name{CES}
\alias{CES}
\title{A function to compute the CES function}
\usage{
CES(K, L, alpha, sigma, eta, nudge = TRUE)
}
\arguments{
\item{K}{Value of input 1.}

\item{L}{Value of input 2.}

\item{alpha}{Share parameter of input 1.}

\item{sigma}{Elasticity of substitution.}

\item{eta}{Returns to scale.}

\item{nudge}{If TRUE, then nudge sigma such that sigma=1 is interpetable.}
}
\value{
CES(K,L)
}
\description{
See Constant elasticity of substitution from Wikipedia
}
