% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMaxs.R
\name{colMaxs}
\alias{colMaxs}
\title{A function to compute the max by col}
\usage{
colMaxs(x, margin = 2)
}
\arguments{
\item{x}{Input array.}

\item{margin}{Margin.}
}
\value{
colMaxs(x, margin)
}
\description{
Simpler helper function.
}
