% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMins.R
\name{colMins}
\alias{colMins}
\title{A function to compute the min by col}
\usage{
colMins(x, margin = 2)
}
\arguments{
\item{x}{Input array.}

\item{margin}{Margin.}
}
\value{
colMins(x, margin)
}
\description{
Simpler helper function.
}
