% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colombia_Data.R
\docType{data}
\name{Colombia_Data}
\alias{Colombia_Data}
\title{Friendship data from a rural Colombian population}
\format{
An object of class \code{"list"}.
}
\source{
\href{https://www.eva.mpg.de/ecology/fieldwork/afrocolombians}{EVA}
}
\usage{
data(Colombia_Data)
}
\description{
Data on friendship nominations paired with covariate data.
}
\examples{
data(Colombia_Data)
}
\references{
Dalla Ragione, A., Ross, CT., and Redhead, DJ.
2022. Friendship and partner choice in rural Colombia.
}
\keyword{datasets}
