% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_posterior.R
\name{convert_posterior}
\alias{convert_posterior}
\title{A function to map NumPyro samples to STRAND format}
\usage{
convert_posterior(posteriors)
}
\arguments{
\item{posteriors}{A NumPyro posterior.}
}
\value{
convert_posterior(posteriors)
}
\description{
This is an internal function, written by Sebastian Sosa.
}
