% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossSection.R
\name{CrossSection}
\alias{CrossSection}
\title{CrossSection class}
\description{
Class for providing access to cross-sectional data.
}
\details{
The CrossSection class provides an interface to cross-sectional data. It
handles data retrieval, validation, value carry-forward and metadata
tracking.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CrossSection$new()}}
\item \href{#method-setCarryForwardValue}{\code{CrossSection$setCarryForwardValue()}}
\item \href{#method-setNAReplaceValue}{\code{CrossSection$setNAReplaceValue()}}
\item \href{#method-setColumnMap}{\code{CrossSection$setColumnMap()}}
\item \href{#method-setRaw}{\code{CrossSection$setRaw()}}
\item \href{#method-getRaw}{\code{CrossSection$getRaw()}}
\item \href{#method-get}{\code{CrossSection$get()}}
\item \href{#method-periodStats}{\code{CrossSection$periodStats()}}
\item \href{#method-print}{\code{CrossSection$print()}}
\item \href{#method-clone}{\code{CrossSection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{CrossSection} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSection$new(carry_forward)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{carry_forward}}{Boolean that indicates whether data in the object
should be carried forward from one period to the next. To carry forward
data is to use data from the past for a security if no data is found
for the present. Suppose data for security X has been retrieved in a
prior call to \code{get()}. Setting \code{carry_forward = TRUE} will
cause the most recent data for X to be used if no records are found for
X in subsequent calls to \code{get()}. If carry-forward values are
supplied using the \code{setCarryForwardValue} method these values are
used instead of the most recent data points.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{CrossSection} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setCarryForwardValue"></a>}}
\if{latex}{\out{\hypertarget{method-setCarryForwardValue}{}}}
\subsection{Method \code{setCarryForwardValue()}}{
Set the carry-forward value(s) for the cross section.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSection$setCarryForwardValue(value_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value_list}}{Object of class \code{list}. Each name-value pair of
the list specifies the name of a column and the value to be used for
that column in records that are carried forward. For example, if
\code{value_list = list(foo = 0)} the value 0 will be used for column
foo in records that are carried forward.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setNAReplaceValue"></a>}}
\if{latex}{\out{\hypertarget{method-setNAReplaceValue}{}}}
\subsection{Method \code{setNAReplaceValue()}}{
Set the value(s) to be used in place of NA observations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSection$setNAReplaceValue(value_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value_list}}{Object of class \code{list}. Each name-value pair of
the list specifies the name of a column and the value to be used in
place of NAs that appear in that column. For example, if
\code{value_list = list(foo = 0)} the value 0 will be used in place of
NA in column foo.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setColumnMap"></a>}}
\if{latex}{\out{\hypertarget{method-setColumnMap}{}}}
\subsection{Method \code{setColumnMap()}}{
Rename columns in input data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSection$setColumnMap(column_map)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column_map}}{Object of class \code{list}. Each name-value pair of
the list represents the new and old column names for the object's data.
For example, if \code{column_map = list(foo = "bar")} values in the raw
data in column bar will appear in the column foo in the data returned
by \code{get()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setRaw"></a>}}
\if{latex}{\out{\hypertarget{method-setRaw}{}}}
\subsection{Method \code{setRaw()}}{
Set raw data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSection$setRaw(raw_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw_data}}{Data frame that contains raw data. Must have a column
\code{date} of class \code{Date}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRaw"></a>}}
\if{latex}{\out{\hypertarget{method-getRaw}{}}}
\subsection{Method \code{getRaw()}}{
Get raw data for a given date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSection$getRaw(date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{Date for which we want data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame of data for \code{date}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
Get cross-section data for a given date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSection$get(date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{Date for which we want data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame of data for \code{date} that includes, if
  \code{carry_forward} was set to \code{TRUE} when creating the object,
  any data that has been carried-forward from previous periods. The
  records carried forward have default values as configured by calls to
  \code{setCarryForwardValue()}. In addition, NA values are replaced
  according the values passed to \code{setNAReplaceValue()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-periodStats"></a>}}
\if{latex}{\out{\hypertarget{method-periodStats}{}}}
\subsection{Method \code{periodStats()}}{
Get overall statistics for the period's data and statistics
  on specific columns (e.g., number of missing values, day-over-day
  correlations, etc.).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSection$periodStats(columns = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{Columns for which to return statistics.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the cross-section object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSection$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
