% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossSectionFromFile.R
\name{CrossSectionFromFile}
\alias{CrossSectionFromFile}
\title{CrossSectionFromFile class}
\description{
Class for providing access to cross-sectional data from binary
  feather files.
}
\details{
The \code{CrossSectionFromFile} class provides an interface to
  cross-sectional data that is contained in binary feather files.
}
\section{Super class}{
\code{\link[strand:CrossSection]{strand::CrossSection}} -> \code{CrossSectionFromFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CrossSectionFromFile$new()}}
\item \href{#method-getRaw}{\code{CrossSectionFromFile$getRaw()}}
\item \href{#method-print}{\code{CrossSectionFromFile$print()}}
\item \href{#method-clone}{\code{CrossSectionFromFile$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="strand" data-topic="CrossSection" data-id="get">}\href{../../strand/html/CrossSection.html#method-get}{\code{strand::CrossSection$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="strand" data-topic="CrossSection" data-id="periodStats">}\href{../../strand/html/CrossSection.html#method-periodStats}{\code{strand::CrossSection$periodStats()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="strand" data-topic="CrossSection" data-id="setCarryForwardValue">}\href{../../strand/html/CrossSection.html#method-setCarryForwardValue}{\code{strand::CrossSection$setCarryForwardValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="strand" data-topic="CrossSection" data-id="setColumnMap">}\href{../../strand/html/CrossSection.html#method-setColumnMap}{\code{strand::CrossSection$setColumnMap()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="strand" data-topic="CrossSection" data-id="setNAReplaceValue">}\href{../../strand/html/CrossSection.html#method-setNAReplaceValue}{\code{strand::CrossSection$setNAReplaceValue()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="strand" data-topic="CrossSection" data-id="setRaw">}\href{../../strand/html/CrossSection.html#method-setRaw}{\code{strand::CrossSection$setRaw()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{CrossSectionFromFile} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSectionFromFile$new(carry_forward, input_dir, input_prefix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{carry_forward}}{Boolean that indicates whether data in the object
should be carried forward from one period to the next.}

\item{\code{input_dir}}{Filesystem (directory) location of input data files.}

\item{\code{input_prefix}}{String prefix for the files that should be loaded
into the cross-section.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{CrossSectionFromFile} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRaw"></a>}}
\if{latex}{\out{\hypertarget{method-getRaw}{}}}
\subsection{Method \code{getRaw()}}{
Get raw data for a given date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSectionFromFile$getRaw(date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{Date for which we want data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame of data for \code{date}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the cross-section object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSectionFromFile$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossSectionFromFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
