% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_set_to_long.R
\name{dyadic_set_to_long}
\alias{dyadic_set_to_long}
\title{A function to map dyadic predictors to a long-form dyad list.}
\usage{
dyadic_set_to_long(X)
}
\arguments{
\item{X}{A 3d-array.}
}
\value{
dyadic_set_to_long(X)
}
\description{
This is an internal function.
}
