% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_strategy_config}
\alias{example_strategy_config}
\title{Load example strategy configuration}
\usage{
example_strategy_config()
}
\value{
An object of class \code{list} that contains the example
  configuration. The list object is the result of loading the package's
  example yaml configuration file \code{application/strategy_config.yaml}.
}
\description{
Loads an example strategy configuration file for use in examples.
}
\examples{

config <- example_strategy_config()
names(config$strategies)
show(config$strategies$strategy_1)

}
