% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bsrm_with_numpyro.R
\name{fit_bsrm_with_numpyro}
\alias{fit_bsrm_with_numpyro}
\title{A NumPyro implementation of the STRAND latent-network SRM}
\usage{
fit_bsrm_with_numpyro(data, mcmc_parameters = mcmc_parameters)
}
\arguments{
\item{data}{A STRAND data object.}

\item{mcmc_parameters}{A parameter list to control NUTS sampling. NumPyro uses the same control set as Stan.}
}
\value{
A NumPyro results object.
}
\description{
A NumPyro implementation of the STRAND latent-network SRM
}
\examples{
\dontrun{
res = fit_bsrm_with_numpyro(data, mcmc_parameters)
}

}
