% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_longitudinal_with_numpyro.R
\name{fit_longitudinal_with_numpyro}
\alias{fit_longitudinal_with_numpyro}
\title{A NumPyro implementation of the STRAND longitudinal latent-network SRM}
\usage{
fit_longitudinal_with_numpyro(data, bandage_penalty, mcmc_parameters)
}
\arguments{
\item{data}{A STRAND data object.}

\item{bandage_penalty}{How tight are the dyadic and longitudinal bindings?}

\item{mcmc_parameters}{A parameter list to control NUTS sampling. NumPyro uses the same control set as Stan.}
}
\value{
A NumPyro results object.
}
\description{
A NumPyro implementation of the STRAND longitudinal latent-network SRM
}
\examples{
\dontrun{
res = fit_longitudinal_with_numpyro(data, bandage_penalty, mcmc_parameters)
}

}
