% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal_set_to_long.R
\name{focal_set_to_long}
\alias{focal_set_to_long}
\title{A function to map focal or sender predictors to a long-form dyad list.}
\usage{
focal_set_to_long(X)
}
\arguments{
\item{X}{A data.frame.}
}
\value{
focal_set_to_long(X)
}
\description{
This is an internal function.
}
