% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodSharing_Data.R
\docType{data}
\name{FoodSharing_Data}
\alias{FoodSharing_Data}
\title{Food and Money sharing data from a rural Colombian population}
\format{
An object of class \code{"list"}.
}
\source{
\href{https://doi.org/10.1098/rsos.192090}{RSOS}
}
\usage{
data(FoodSharing_Data)
}
\description{
Self-report data on transfers, double-sampled, paired with covariate data.
}
\examples{
data(FoodSharing_Data)
}
\references{
Anne C. Pisor, Matthew M. Gervais, Benjamin G. Purzycki, and Cody T. Ross. 2020.
Preferences and constraints: the value of economic games for studying human behaviour
Royal Society Open Science. Volume 7, Issue 6.
}
\keyword{datasets}
