% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_structured_correlation_matrix.R
\name{generate_structured_correlation_matrix}
\alias{generate_structured_correlation_matrix}
\title{A function to generate correlation matrices with special inner symmetry}
\usage{
generate_structured_correlation_matrix(
  data,
  eta,
  mode,
  setting,
  mcmc_parameters
)
}
\arguments{
\item{data}{A data list.}

\item{eta}{Prior on the LKJ Cholesky factor.}

\item{mode}{Which method to use? "cholesky" or "l2norm".}

\item{setting}{Which matrix to create? "multiplex_dyadic_reciprocity", "longitudinal_dyadic_reciprocity", or "longitudinal_generalized_reciprocity".}

\item{mcmc_parameters}{A list of Stan parameters that often need to be tuned.}
}
\value{
A Stan object.
}
\description{
This is an internal function.
}
