% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDI.R
\name{HPDI}
\alias{HPDI}
\title{A function to get Highest Posterior Density Interval}
\usage{
HPDI(samples, prob = 0.89)
}
\arguments{
\item{samples}{A vector of samples to summarize.}

\item{prob}{Interval range.}
}
\value{
A vector of intervals.
}
\description{
This is a simple helper function to get density intervals. See McElreaths Rethinking for details.
}
