% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit.R
\name{inv_logit}
\alias{inv_logit}
\title{A function to take inverse logit}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{value}
}
\value{
inverse_logit(x)
}
\description{
See McElreaths Rethinking package for details
}
