% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit_shifted.R
\name{inv_logit_shifted}
\alias{inv_logit_shifted}
\title{A function to take inverse logit shifted}
\usage{
inv_logit_shifted(x)
}
\arguments{
\item{x}{value}
}
\value{
inv_logit_shifted(x)
}
\description{
Map x from (0, inf) to (0, 1)
}
