% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_error.R
\name{is_error}
\alias{is_error}
\title{A function to check for errors}
\usage{
is_error(expr, tell = FALSE, force = FALSE)
}
\arguments{
\item{expr}{expression}

\item{tell}{Logical: Should the error message be printed?}

\item{force}{Logical: Should an error be returned if the expression is not an error?}
}
\value{
Boolean for error being thrown.
}
\description{
Taken from berryFunctions by Berry Boessenkool.
See https://github.com/brry/berryFunctions for details.
}
