% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_numeric_not_binary.R
\name{is_numeric_not_binary}
\alias{is_numeric_not_binary}
\title{A function to check the data type}
\usage{
is_numeric_not_binary(x)
}
\arguments{
\item{x}{value}
}
\value{
is_numeric_not_binary(x)
}
\description{
A function to check the data type
}
