% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jax_summary.R
\name{jax_summary}
\alias{jax_summary}
\title{A function to compute summary stats from a JAX object}
\usage{
jax_summary(X, n_iter, n_chains)
}
\arguments{
\item{X}{A set of sample to summarize.}

\item{n_iter}{Iterations.}

\item{n_chains}{Chains.}
}
\value{
A table of MCMC summaries.
}
\description{
This is a small helper function for JAX ESS and Rhat checking.
}
