% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_to_dyadic_set.R
\name{long_to_dyadic_set}
\alias{long_to_dyadic_set}
\title{A function to map a long-form dyad list to a matrix style.}
\usage{
long_to_dyadic_set(X, N_id)
}
\arguments{
\item{X}{A set of S samples from an array of S x Ndyad x 2.}

\item{N_id}{Rows in the dyad list.}
}
\value{
long_to_dyadic_set(X, N_id)
}
\description{
This is an internal function.
}
