% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_plot_b.R
\name{longitudinal_plot_b}
\alias{longitudinal_plot_b}
\title{Plot longitudinal random effects - dyadic reciprocity}
\usage{
longitudinal_plot_b(
  input,
  results,
  plot = TRUE,
  save_plot = NULL,
  height = 6,
  width = 6,
  palette = NULL
)
}
\arguments{
\item{input}{A STRAND model object, obtained by fitting a longitudinal combined stochastic block and social relations model.}

\item{results}{A results data.frame or matrix with colnames: c("Variable", "Layer", "Target" , "Base" , "Median", "L", "H", "Mean","SD","LayerNumeric", "P")}

\item{plot}{Should a plot be displayed?}

\item{save_plot}{Should a plot be exported to working directory? If so, set save_plot="desired_filename.pdf".}

\item{height}{Height of exported plot.}

\item{width}{Width of exported plot.}

\item{palette}{Override the default palette with a 3-vector of color codes.}
}
\value{
A figure or tabluar data to make a figure.
}
\description{
Plot longitudinal random effects - dyadic reciprocity
}
