% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_plot_c.R
\name{longitudinal_plot_c}
\alias{longitudinal_plot_c}
\title{Plot longitudinal random effects - dyadic reciprocity}
\usage{
longitudinal_plot_c(
  input,
  parameters,
  type,
  normalized = FALSE,
  HPDI = 0.9,
  plot = TRUE,
  export_as_table = FALSE,
  save_plot = NULL,
  height = 6,
  width = 6,
  palette = NULL
)
}
\arguments{
\item{input}{A STRAND model object, obtained by fitting a longitudinal combined stochastic block and social relations model.}

\item{parameters}{Name of the parameter as included in the function call. E.g., ~ Age + Sex... Then: parameters=c("Age", "Age", "Sex", "Sex")}

\item{type}{For each parameter, indicate its type: "focal", "target", or "dyadic". And, from the above: type = =c("focal", "target", "focal", "target")}

\item{normalized}{Should effects be normalized?}

\item{HPDI}{Highest Posterior Density Interval. Ranges in (0,1).}

\item{plot}{Should a plot be displayed?}

\item{export_as_table}{Should the tabular data rather than a ggplot object be returned?}

\item{save_plot}{Should a plot be exported to working directory? If so, set save_plot="desired_filename.pdf".}

\item{height}{Height of exported plot.}

\item{width}{Width of exported plot.}

\item{palette}{Override the default palette with a 3-vector of color codes.}
}
\value{
A figure or tabluar data to make a figure.
}
\description{
Plot longitudinal random effects - dyadic reciprocity
}
\examples{
\dontrun{
res = longitudinal_plot_c(input = fit, parameters, type)
}

}
