% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_plot_g.R
\name{longitudinal_plot_g}
\alias{longitudinal_plot_g}
\title{Plot multiplex random effects - generalized reciprocity}
\usage{
longitudinal_plot_g(
  input,
  HPDI = 0.9,
  plot = TRUE,
  export_as_table = FALSE,
  save_plot = NULL,
  height = 6,
  width = 6,
  palette = NULL
)
}
\arguments{
\item{input}{A STRAND model object, obtained by fitting a multiplex combined stochastic block and social relations model.}

\item{HPDI}{Highest Posterior Density Interval. Ranges in (0,1).}

\item{plot}{Should a plot be displayed?}

\item{export_as_table}{Should the tabular data rather than a ggplot object be returned?}

\item{save_plot}{Should a plot be exported to working directory? If so, set save_plot="desired_filename.pdf".}

\item{height}{Height of exported plot.}

\item{width}{Width of exported plot.}

\item{palette}{Override the default palette with a 3-vector of color codes.}
}
\value{
A figure or tabluar data to make a figure.
}
\description{
Plot multiplex random effects - generalized reciprocity
}
\examples{
\dontrun{
res = multiplex_plot_g(input = fit)
}

}
