% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dyadic_edgelist.R
\name{make_dyadic_edgelist}
\alias{make_dyadic_edgelist}
\title{A function to create a long-form dyad list.}
\usage{
make_dyadic_edgelist(N_id)
}
\arguments{
\item{N_id}{A count of nodes.}
}
\value{
make_dyadic_edgelist(N_id)
}
\description{
This is an internal function.
}
