% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{make_ft}
\alias{make_ft}
\title{Make Basic Flextable}
\usage{
make_ft(x, title = NULL, col_names = NULL, hlines = "all")
}
\arguments{
\item{x}{The data.frame to use for flextable}

\item{title}{The string to use as the table title}

\item{col_names}{A character vector of preferred column names for flextable. 
Length of character vector must be equal to the number of columns. Defaults 
to NULL, in which case the column names of x are used in the flextable.}

\item{hlines}{The row numbers to draw horizontal lines beneath. Defaults to 
"all", can be "all", "none", or a numeric vector.}
}
\value{
A flextable object with the argued formatting
}
\description{
Make a flextable with preferred formatting
}
