% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{make_priors}
\alias{make_priors}
\title{A function to create priors for STRAND models}
\usage{
make_priors(priors_to_change = NULL, include_rownames = FALSE)
}
\arguments{
\item{priors_to_change}{A named vector of priors to change. To get parameter names, run: make_priors(include_rownames=TRUE). Values must be 2-vectors. If the prior has only one parameter, inlcude 0 as the second element.}

\item{include_rownames}{Should rownames be printed?}
}
\value{
A STRAND priors matrix.
}
\description{
This function allows users to change the numerical values of priors (but not their distributions). Users should only change these values if they have read and understood the source code.
}
\examples{
\dontrun{
make_priors(priors_to_change=list(sr_L=5),include_rownames=TRUE)
make_priors()
}

}
