% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_mcmc_parameters.R
\name{merge_mcmc_parameters}
\alias{merge_mcmc_parameters}
\title{A function to update mcmc control parameters}
\usage{
merge_mcmc_parameters(user_params = list())
}
\arguments{
\item{user_params}{User-supplied updates}
}
\value{
a list of parameters
}
\description{
A function to update mcmc control parameters
}
