% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix_strand.R
\name{model_matrix_strand}
\alias{model_matrix_strand}
\title{A wrapper function for model.matrix that reports more helpful errors.}
\usage{
model_matrix_strand(x, d, name)
}
\arguments{
\item{x}{Model equation.}

\item{d}{Data object.}

\item{name}{Name of equation for error message.}
}
\value{
Boolean for error being thrown.
}
\description{
A wrapper function for model.matrix that reports more helpful errors.
}
