% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{normalize}
\alias{normalize}
\title{Transforms an input variable to a normal distribution.}
\usage{
normalize(
  x,
  in_var = NULL,
  within_var = NULL,
  out_sd = NULL,
  adj_var = NULL,
  loops = 3
)
}
\arguments{
\item{x}{data frame, containing all the necessary input data.}

\item{in_var}{character vector, the name of the variable to be adjusted.}

\item{within_var}{character vector, variables within which we 
will manipulate the distribution of \code{in_var} to be N(0,1).}

\item{out_sd}{a list of lists which specify the output standard deviations to
use for different levels of specified variables. The variables here have no
necessary connection with the variables in \code{within_var}.}

\item{adj_var}{character vector, the name of the variable(s) to which we 
want \code{in_var} to be neutral. For now, neutral means uncorrelated 
with. In later versions, we hope this will mean no decile/quintile exposure
to.}

\item{loops}{numeric, number of iterations to run.}
}
\value{
a numeric vector equal in length to \code{in_var}.
}
\description{
\code{normalize} transforms an \code{in_var} to a normal distribution using 
ranks. Output is also adjusted in accordance with the arguments passed in using 
\code{within_var}, \code{out_sd} and \code{adj_var}. It is, generally, 
impossible for the result to meet all these requirements exactly. See
comments in the code for details.
}
