% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_longitudinal_data_missings.R
\name{parse_longitudinal_data_missings}
\alias{parse_longitudinal_data_missings}
\title{An internal function to parse a list of STRAND data objects into a longitudinal object and then apply regression models.}
\usage{
parse_longitudinal_data_missings(
  data,
  block_regression,
  focal_regression,
  target_regression,
  dyad_regression
)
}
\arguments{
\item{data}{A data objects of class STRAND, prepared using the make_strand_data() function. The data object must include all covariates used in the formulas of the fit_longitudinal_model function.}

\item{block_regression}{A formula for the block-level predictors. This should be specified as in lm(), e.g.: ~ Ethnicity + Sex. Dont use interactions, however.}

\item{focal_regression}{A formula for the predictors of out-degree (i.e., focal effects, or the effects of individual covariates on outgoing ties). This should be specified as in lm(), e.g.: ~ Age * Education}

\item{target_regression}{A formula for the predictors of in-degree (i.e., target effects, or the effects of individual covariates on incoming ties). This should be specified as in lm(), e.g.: ~ Age * Education}

\item{dyad_regression}{A formula for the predictors of dyadic relationships. This should be specified as in lm(), e.g.: ~ Kinship + Friendship}
}
\value{
A STRAND data object.
}
\description{
An internal function to parse a list of STRAND data objects into a longitudinal object and then apply regression models.
}
