% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diffusion.R
\name{plot_diffusion}
\alias{plot_diffusion}
\title{A function to simulate a network-based diffusion proccess using the STRAND framework}
\usage{
plot_diffusion(long_data, palette = c("#13667B", "#35ABC0"))
}
\arguments{
\item{long_data}{A list of data objects of class STRAND prepared using the make_strand_data() function. The data objects must include all covariates and trait diffusion data used in the formulas listed below.}

\item{palette}{A 2-vector of plot colors. First element is for the low value of scale, second is for the high value.}
}
\value{
A ggplot2 object.
}
\description{
This function allows users to simulate data using a NBDA model. The user must supply a list of STRAND data objects, a set of parameters,
and a series of formulas following standard lm() style syntax.
}
\examples{
\dontrun{
plot_diffusion(long_data)
}

}
