% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{Portfolio}
\alias{Portfolio}
\title{Portfolio class}
\description{
The portfolio class represents a collection of positions.
}
\details{
This class wraps a tabular data structure that contains position
information. Methods are available for returning this information in long or
wide format.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Portfolio$new()}}
\item \href{#method-getPositions}{\code{Portfolio$getPositions()}}
\item \href{#method-setPositions}{\code{Portfolio$setPositions()}}
\item \href{#method-removePositions}{\code{Portfolio$removePositions()}}
\item \href{#method-getConsolidatedPositions}{\code{Portfolio$getConsolidatedPositions()}}
\item \href{#method-getShareColumns}{\code{Portfolio$getShareColumns()}}
\item \href{#method-applyAdjustmentRatio}{\code{Portfolio$applyAdjustmentRatio()}}
\item \href{#method-clone}{\code{Portfolio$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Portfolio} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$new(strategy_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{strategy_names}}{An object of class character that contains the
strategy names for this portfolio object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Portfolio} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPositions"></a>}}
\if{latex}{\out{\hypertarget{method-getPositions}{}}}
\subsection{Method \code{getPositions()}}{
Position access.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$getPositions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame that contains the raw position information for this
  portfolio. The data frame contains the following columns: id
  (character), strategy (character), int_shares (integer) for the number
  of shares that net down with shares in other strategies, and
  ext_shares (integer) for the number of shares that do not net down with
  any other strategy's positions. The sum of the values in the ext_share
  column represent the number of shares held at the joint level (i.e.,
  that don't cancel out with other positions).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPositions"></a>}}
\if{latex}{\out{\hypertarget{method-setPositions}{}}}
\subsection{Method \code{setPositions()}}{
Replace this portfolio's position information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$setPositions(positions_df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{positions_df}}{Data frame that contains the portfolio's low-level
position information. The data frame must contain the following
columns: id (character), strategy (character), int_shares (integer),
ext_shares (integer). No NAs are allowed. All values in the strategy
column must be one of the strategy names for this portfolio.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-removePositions"></a>}}
\if{latex}{\out{\hypertarget{method-removePositions}{}}}
\subsection{Method \code{removePositions()}}{
Remove positions from the portfolio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$removePositions(ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{Character vector that contains the security ids to remove.
Positions in all strategies for the supplied ids are removed from the
portfolio.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConsolidatedPositions"></a>}}
\if{latex}{\out{\hypertarget{method-getConsolidatedPositions}{}}}
\subsection{Method \code{getConsolidatedPositions()}}{
Access to positions in consolidated wide format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$getConsolidatedPositions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame of consolidated (external + internal) share values
  in wide format. The columns in the data frame returned by this method
  are id, strategy, and shares_{s} for each strategy s.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getShareColumns"></a>}}
\if{latex}{\out{\hypertarget{method-getShareColumns}{}}}
\subsection{Method \code{getShareColumns()}}{
Helper for extracting strategy share columns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$getShareColumns()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of column names that matches the strategy columns in the
  data frame returned by \code{getConsolidatedPositions}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-applyAdjustmentRatio"></a>}}
\if{latex}{\out{\hypertarget{method-applyAdjustmentRatio}{}}}
\subsection{Method \code{applyAdjustmentRatio()}}{
Adjust portfolio share values by an adjustment ratio. The
  adjustment ratio is the number of old shares per new share. So for a
  2:1 split, the adjustment ratio is 0.5.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$applyAdjustmentRatio(adjustment_df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{adjustment_df}}{Data frame that contains adjustment ratios to apply.
The data frame must include two columns: id (character) and
adjustment_ratio (numeric). There must be an entry in adjustment_df for
each security in the portfolio (this requirement will likely be
loosened in future versions).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
