% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_block_parameters.R
\name{process_block_parameters}
\alias{process_block_parameters}
\title{Calculate the contrast between two block estimates for each layer}
\usage{
process_block_parameters(input, focal, base, HPDI = 0.9)
}
\arguments{
\item{input}{A STRAND model object}

\item{focal}{The outcome of interest, will be of the form "X to Y", where X and Y are factor levels from the block model.}

\item{base}{The base case, will be of the form "X to Y", where X and Y are factor levels from the block model.}

\item{HPDI}{Highest Posterior Density Interval. Ranges in (0,1).}
}
\value{
A matrix of output.
}
\description{
This is a helper function to compute contrast between samples
}
