% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_missings.R
\name{process_missings}
\alias{process_missings}
\title{A function to compute relevant data on missings for Stan imputation}
\usage{
process_missings(data)
}
\arguments{
\item{data}{A data object from make_strand_data after its proccessed by a model function call.}
}
\value{
process_missings(data)
}
\description{
Simpler helper function.
}
