% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbern.R
\name{rbern}
\alias{rbern}
\title{A function to simulate Bernoulli data}
\usage{
rbern(n, prob = 0.5)
}
\arguments{
\item{n}{Number of samples}

\item{prob}{Probability}
}
\value{
Binary samples
}
\description{
See McElreaths Rethinking package for details
}
