% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdirichlet.R
\name{rdirichlet}
\alias{rdirichlet}
\title{A function to simulate Dirichlet data}
\usage{
rdirichlet(n, mu)
}
\arguments{
\item{n}{number of samples}

\item{mu}{mean vector}
}
\value{
samples
}
\description{
See https://github.com/dkahle/dirichlet/blob/master/R/dirichlet.R for details
}
