% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_jax_arrays.R
\name{reshape_jax_arrays}
\alias{reshape_jax_arrays}
\title{A function to reshape a JAX samples object}
\usage{
reshape_jax_arrays(samps, n_iter, n_chains)
}
\arguments{
\item{samps}{A set of samples to summarize.}

\item{n_iter}{Iterations.}

\item{n_chains}{Chains.}
}
\value{
A reshaped array that posterior can work with.
}
\description{
This is a small helper function for JAX ESS and Rhat checking.
}
