% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RICH_Data.R
\docType{data}
\name{RICH_Data}
\alias{RICH_Data}
\title{Data on RICH games from Colombia}
\format{
An object of class \code{"list"}. N individuals = 93. The list contains three outcome matrices (of dimension 93 x 93) containing data on giving, taking, and reducing behaviours.
}
\source{
\href{http://dx.doi.org/}{Working paper}
}
\usage{
data(RICH_Data)
}
\description{
Data on network-structured economic games with additional covariate data. These data have been rescaled, and some missing values were imputed using the mode or median. See the reference to the original study below for full methodological information.
}
\examples{
data(RICH_Data)
}
\references{
Ross and Pisor, 2024.
Perceived inequality and variability in the expression of parochial altruism
}
\keyword{datasets}
