% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlkjcorr.R
\name{rlkjcorr}
\alias{rlkjcorr}
\title{A function to generate rlkjcorr matrix}
\usage{
rlkjcorr(n, K, eta = 1)
}
\arguments{
\item{n}{number of samples}

\item{K}{dimension}

\item{eta}{Prior density controller}
}
\value{
Rho
}
\description{
See McElreaths Rethinking package for details
}
