% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm2.R
\name{rmvnorm2}
\alias{rmvnorm2}
\title{A function to simulate MV Normal data}
\usage{
rmvnorm2(
  n,
  Mu = rep(0, length(sigma)),
  sigma = rep(1, length(Mu)),
  Rho = diag(length(Mu)),
  method = "chol"
)
}
\arguments{
\item{n}{Number of samples}

\item{Mu}{Mean vector}

\item{sigma}{variance vector}

\item{Rho}{Correlation matrix}

\item{method}{See details in Rethinking.}
}
\value{
MV normal samples
}
\description{
See McElreaths Rethinking package for details
}
