% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_longitudinal_checks.R
\name{run_longitudinal_checks}
\alias{run_longitudinal_checks}
\title{An internal function to check that a list of STRAND data objects can be merged into a longitudinal object}
\usage{
run_longitudinal_checks(data, pass)
}
\arguments{
\item{data}{A list of data objects of class STRAND prepared using the make_strand_data() function. The data object must include all covariates used in the formulas of the fit_longitudinal_model() function.}

\item{pass}{1 or 2. First pass checks raw data, second pass checks some computed values also.}
}
\value{
Returns an error if objects do not match.
}
\description{
An internal function to check that a list of STRAND data objects can be merged into a longitudinal object
}
