% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_pricing}
\alias{sample_pricing}
\title{Sample pricing data for examples and testing}
\format{
A data frame with 31500 rows and 8 variables:
\describe{
  \item{id}{security identifier}
  \item{date}{pricing date}
  \item{price_unadj}{the unadjusted price of the security}
  \item{prior_close_unadj}{the unadjusted prior closing price of the security}
  \item{dividend_unadj}{the dividend for the security on an unadjusted basis}
  \item{distribution_unadj}{the distribution (e.g., spin-off) for the security on an unadjusted basis}
  \item{volume}{trading volume for the security, in shares}
  \item{adjustment_ratio}{the adjustment ratio for the security}
}
}
\usage{
sample_pricing
}
\description{
A dataset containing sample (fake) pricing data for 500 securities and
63 weekdays, from 2019-01-02 to 2019-03-29.
}
\keyword{datasets}
