% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{show_best_worst}
\alias{show_best_worst}
\title{Show Best/Worst Performers}
\usage{
show_best_worst(sim)
}
\arguments{
\item{sim}{A Simulation object to show the best and worst performers for}
}
\description{
Build a flextable object showing a Simulation's best and worst 
performers
}
