% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{show_config}
\alias{show_config}
\title{Show Strategy Configuration}
\usage{
show_config(sim)
}
\arguments{
\item{sim}{A Simulation object to show the configuration for}
}
\description{
Build a flextable object showing a Simulation's configuration
}
