% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{show_constraints}
\alias{show_constraints}
\title{Show Strategy Constraints}
\usage{
show_constraints(sim)
}
\arguments{
\item{sim}{A Simulation object to show the configuration for}
}
\description{
Build a flextable object showing a Simulation's risk constraints
}
