% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{show_monthly_returns}
\alias{show_monthly_returns}
\title{Show monthly returns}
\usage{
show_monthly_returns(sim)
}
\arguments{
\item{sim}{A Simulation object with results to display}
}
\description{
Build a flextable object that shows a simulation's return by
  month by formatting the output of `Simulation$overallReturnsByMonthDf`.
}
