% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{show_stats}
\alias{show_stats}
\title{Show Overall Stats Table}
\usage{
show_stats(sim)
}
\arguments{
\item{sim}{A Simulation object to show the statistics for}
}
\description{
Build a flextable object showing a Simulation's overall statistics
}
