% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimResult.R
\name{SimResult}
\alias{SimResult}
\title{SimResult class}
\description{
Class for storing simulation results.
}
\details{
The \code{SimResult} class wraps simulation results and provides methods for
accessing, summarizing, and plotting those results.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SimResult$new()}}
\item \href{#method-getConfig}{\code{SimResult$getConfig()}}
\item \href{#method-saveSimSummary}{\code{SimResult$saveSimSummary()}}
\item \href{#method-saveSimDetail}{\code{SimResult$saveSimDetail()}}
\item \href{#method-saveInputStats}{\code{SimResult$saveInputStats()}}
\item \href{#method-saveLooseningInfo}{\code{SimResult$saveLooseningInfo()}}
\item \href{#method-saveOptimizationSummary}{\code{SimResult$saveOptimizationSummary()}}
\item \href{#method-saveExposures}{\code{SimResult$saveExposures()}}
\item \href{#method-saveDelistings}{\code{SimResult$saveDelistings()}}
\item \href{#method-getSimSummary}{\code{SimResult$getSimSummary()}}
\item \href{#method-getSimDetail}{\code{SimResult$getSimDetail()}}
\item \href{#method-getPositionSummary}{\code{SimResult$getPositionSummary()}}
\item \href{#method-getInputStats}{\code{SimResult$getInputStats()}}
\item \href{#method-getLooseningInfo}{\code{SimResult$getLooseningInfo()}}
\item \href{#method-getOptimizationSummary}{\code{SimResult$getOptimizationSummary()}}
\item \href{#method-getExposures}{\code{SimResult$getExposures()}}
\item \href{#method-getDelistings}{\code{SimResult$getDelistings()}}
\item \href{#method-getSingleStrategySummaryDf}{\code{SimResult$getSingleStrategySummaryDf()}}
\item \href{#method-plotPerformance}{\code{SimResult$plotPerformance()}}
\item \href{#method-plotMarketValue}{\code{SimResult$plotMarketValue()}}
\item \href{#method-plotCategoryExposure}{\code{SimResult$plotCategoryExposure()}}
\item \href{#method-plotFactorExposure}{\code{SimResult$plotFactorExposure()}}
\item \href{#method-plotNumPositions}{\code{SimResult$plotNumPositions()}}
\item \href{#method-overallStatsDf}{\code{SimResult$overallStatsDf()}}
\item \href{#method-print}{\code{SimResult$print()}}
\item \href{#method-writeFeather}{\code{SimResult$writeFeather()}}
\item \href{#method-readFeather}{\code{SimResult$readFeather()}}
\item \href{#method-clone}{\code{SimResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SimResult} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$new(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{The list of configuration information used to run the
simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SimResult} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getConfig"></a>}}
\if{latex}{\out{\hypertarget{method-getConfig}{}}}
\subsection{Method \code{getConfig()}}{
Get the object's configuration information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$getConfig()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Object of class \code{list} that contains the simulation's
  configuration information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveSimSummary"></a>}}
\if{latex}{\out{\hypertarget{method-saveSimSummary}{}}}
\subsection{Method \code{saveSimSummary()}}{
Save summary information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$saveSimSummary(period, data_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{period}}{Period to which the data pertains.}

\item{\code{data_obj}}{Data frame to save.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveSimDetail"></a>}}
\if{latex}{\out{\hypertarget{method-saveSimDetail}{}}}
\subsection{Method \code{saveSimDetail()}}{
Save detail (position-level) information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$saveSimDetail(period, data_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{period}}{Period to which the data pertains.}

\item{\code{data_obj}}{Data frame to save.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveInputStats"></a>}}
\if{latex}{\out{\hypertarget{method-saveInputStats}{}}}
\subsection{Method \code{saveInputStats()}}{
Save input statistics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$saveInputStats(period, data_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{period}}{Period to which the data pertains.}

\item{\code{data_obj}}{Data frame to save.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveLooseningInfo"></a>}}
\if{latex}{\out{\hypertarget{method-saveLooseningInfo}{}}}
\subsection{Method \code{saveLooseningInfo()}}{
Save loosening information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$saveLooseningInfo(period, data_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{period}}{Period to which the data pertains.}

\item{\code{data_obj}}{Data frame to save.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveOptimizationSummary"></a>}}
\if{latex}{\out{\hypertarget{method-saveOptimizationSummary}{}}}
\subsection{Method \code{saveOptimizationSummary()}}{
Save optimization summary information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$saveOptimizationSummary(period, data_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{period}}{Period to which the data pertains.}

\item{\code{data_obj}}{Data frame to save.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveExposures"></a>}}
\if{latex}{\out{\hypertarget{method-saveExposures}{}}}
\subsection{Method \code{saveExposures()}}{
Save exposure information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$saveExposures(period, data_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{period}}{Period to which the data pertains.}

\item{\code{data_obj}}{Data frame to save.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveDelistings"></a>}}
\if{latex}{\out{\hypertarget{method-saveDelistings}{}}}
\subsection{Method \code{saveDelistings()}}{
Save information on positions removed due to delisting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$saveDelistings(period, data_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{period}}{Period to which the data pertains.}

\item{\code{data_obj}}{Data frame to save.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSimSummary"></a>}}
\if{latex}{\out{\hypertarget{method-getSimSummary}{}}}
\subsection{Method \code{getSimSummary()}}{
Get summary information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$getSimSummary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame that contains all summary data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSimDetail"></a>}}
\if{latex}{\out{\hypertarget{method-getSimDetail}{}}}
\subsection{Method \code{getSimDetail()}}{
Get detail information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$getSimDetail(
  sim_date = NULL,
  strategy_name = NULL,
  security_id = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sim_date}}{Vector of length 1 of class Date or character that
specifies the period for which to get detail information. If
\code{NULL} then data from all periods is returned. Note that as
opposed to filtering on a date column in the detail data itself, the
period index (i.e., the list element name) is used when extracting data
for a single period. This is the same value that is supplied as the
\code{period} parameter when calling \code{saveSimDetail()}. Defaults
to \code{NULL}.}

\item{\code{strategy_name}}{Character vector of length 1 that specifies the
strategy for which to get detail data. If \code{NULL} data for all
strategies is returned. Defaults to \code{NULL}.}

\item{\code{security_id}}{Character vector of length 1 that specifies the
security for which to get detail data. If \code{NULL} data for all
securities is returned. Defaults to \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame that contains detail data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPositionSummary"></a>}}
\if{latex}{\out{\hypertarget{method-getPositionSummary}{}}}
\subsection{Method \code{getPositionSummary()}}{
Get summary information by security. This method can be
  used, for example, to calculate the biggest winners and losers over the
  course of the simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$getPositionSummary(strategy_name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{strategy_name}}{Character vector of length 1 that specifies the
strategy for which to get detail data. If \code{NULL} data for all
strategies is returned. Defaults to \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame of summary information aggregated by security.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getInputStats"></a>}}
\if{latex}{\out{\hypertarget{method-getInputStats}{}}}
\subsection{Method \code{getInputStats()}}{
Get input statistics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$getInputStats()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame that contains all input statistics.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLooseningInfo"></a>}}
\if{latex}{\out{\hypertarget{method-getLooseningInfo}{}}}
\subsection{Method \code{getLooseningInfo()}}{
Get loosening information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$getLooseningInfo()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame that contains all loosening information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOptimizationSummary"></a>}}
\if{latex}{\out{\hypertarget{method-getOptimizationSummary}{}}}
\subsection{Method \code{getOptimizationSummary()}}{
Get optimization summary information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$getOptimizationSummary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame that contains all optimization summary information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getExposures"></a>}}
\if{latex}{\out{\hypertarget{method-getExposures}{}}}
\subsection{Method \code{getExposures()}}{
Get exposure information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$getExposures()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame that contains all exposure information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDelistings"></a>}}
\if{latex}{\out{\hypertarget{method-getDelistings}{}}}
\subsection{Method \code{getDelistings()}}{
Get information on positions removed due to delisting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$getDelistings()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame that contains all delisting information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSingleStrategySummaryDf"></a>}}
\if{latex}{\out{\hypertarget{method-getSingleStrategySummaryDf}{}}}
\subsection{Method \code{getSingleStrategySummaryDf()}}{
Get summary information for a single strategy suitable for
  plotting input.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$getSingleStrategySummaryDf(strategy_name, include_zero_row = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{strategy_name}}{Strategy for which to return summary data.}

\item{\code{include_zero_row}}{Logical flag indicatiing whether to prepend a row
to the summary data with starting values at zero. Defaults to \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame that contains summary information for the desired
  strategy, as well as columns for cumulative net and gross total return,
  calculated as pnl divided by ending gross market value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPerformance"></a>}}
\if{latex}{\out{\hypertarget{method-plotPerformance}{}}}
\subsection{Method \code{plotPerformance()}}{
Plot cumulative gross and net return by date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$plotPerformance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotMarketValue"></a>}}
\if{latex}{\out{\hypertarget{method-plotMarketValue}{}}}
\subsection{Method \code{plotMarketValue()}}{
Plot total gross, long, short, and net market value by date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$plotMarketValue()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotCategoryExposure"></a>}}
\if{latex}{\out{\hypertarget{method-plotCategoryExposure}{}}}
\subsection{Method \code{plotCategoryExposure()}}{
Plot exposure to all levels in a category by date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$plotCategoryExposure(in_var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_var}}{Category for which exposures are plotted.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFactorExposure"></a>}}
\if{latex}{\out{\hypertarget{method-plotFactorExposure}{}}}
\subsection{Method \code{plotFactorExposure()}}{
Plot exposure to factors by date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$plotFactorExposure(in_var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_var}}{Factors for which exposures are plotted.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotNumPositions"></a>}}
\if{latex}{\out{\hypertarget{method-plotNumPositions}{}}}
\subsection{Method \code{plotNumPositions()}}{
Plot number of long and short positions by date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$plotNumPositions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-overallStatsDf"></a>}}
\if{latex}{\out{\hypertarget{method-overallStatsDf}{}}}
\subsection{Method \code{overallStatsDf()}}{
Calculate overall simulation summary statistics, such as
  total P&L, Sharpe, average market values and counts, etc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$overallStatsDf()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame that contains summary statistics, suitable for
  reporting.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print overall simulation statistics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-writeFeather"></a>}}
\if{latex}{\out{\hypertarget{method-writeFeather}{}}}
\subsection{Method \code{writeFeather()}}{
Write the data in the object to feather files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$writeFeather(out_loc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out_loc}}{Directory in which output files should be created.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-readFeather"></a>}}
\if{latex}{\out{\hypertarget{method-readFeather}{}}}
\subsection{Method \code{readFeather()}}{
Load files created with \code{writeFeather} into the object.
  Note that because detail data is not re-split by period, it will not be
  possible to use the \code{sim_date} parameter when calling
  \code{getSimDetail} on the populated object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$readFeather(in_loc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_loc}}{Directory that contains files to be loaded.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
