% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sbm_network.R
\name{simulate_sbm_network}
\alias{simulate_sbm_network}
\title{A function to simulate single layer directed networks with a stochastic block structure}
\usage{
simulate_sbm_network(
  N_id = 90,
  N_groups = 3,
  group_probs = c(0.2, 0.5, 0.3),
  B = NULL,
  in_block = 0.02,
  out_block = 0.01,
  sr_mu = c(0, 0),
  sr_sigma = c(1, 1),
  sr_rho = 0.3,
  dr_mu = c(0, 0),
  dr_sigma = 1,
  dr_rho = 0.7,
  individual_predictors = NULL,
  dyadic_predictors = NULL,
  individual_effects = NULL,
  dyadic_effects = NULL
)
}
\arguments{
\item{N_id}{Number of individuals.}

\item{N_groups}{Number of groups (i.e., blocks).}

\item{group_probs}{A vector of the probabilities of individuals being in each group.}

\item{B}{Tie probabilities between group blocks. If B is null, then in_block and out_block can be specified to create B.}

\item{in_block}{Tie probabilities between members of the same group. This is overridden by a non-NULL B parameter.}

\item{out_block}{Tie probabilities between memebers of different groups. This is overridden by a non-NULL B parameter.}

\item{individual_predictors}{An N_id by N_individual_parameters matrix of covariates.}

\item{dyadic_predictors}{An N_id by N_id by N_dyadic_parameters array of covariates.}

\item{individual_effects}{A 2 by N_individual_parameters matrix of slopes. The first row gives effects of focal characteristics (on out-degree). 
The second row gives effects of target characteristics (on in-degree).}

\item{dyadic_effects}{An N_dyadic_parameters vector of slopes.}
}
\value{
A list of 5 objects: network (an adjacency matrix of binary outcomes), tie_strength (an adjacency matrix with probability weights), 
group_ids (a vector of length N_id, giving the group of each individual), individual_predictors (the supplied covariate data is saved along with the network data), 
and dyadic_predictors (the supplied covariate data is saved along with the network data).
}
\description{
This is a function to simulate single layer network data with a stochastic block structure.
}
\examples{
\dontrun{
B = diag(3)*0.04
B[1,3] = 0.010
B[3,2] = 0.007
B[1,1] = 0.07

A = simulate_sbm_network(N_id = 100, B=B, individual_predictor=matrix(rnorm(100, 0, 1), nrow=100,ncol=1), individual_effects=matrix(c(1.2, 0.5),ncol=1,nrow=2))
Net = graph_from_adjacency_matrix(A$network, mode = c("directed"))
V(Net)$color = c("turquoise4","gray13", "goldenrod3")[A$group_ids]

plot(Net, edge.arrow.size =0.1, edge.curved = 0.3, vertex.label=NA, vertex.size = 5)
}

}
