% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sbm_network_bpb.R
\name{simulate_sbm_network_bpb}
\alias{simulate_sbm_network_bpb}
\title{A function to simulate single layer directed networks with a stochastic block structure}
\usage{
simulate_sbm_network_bpb(
  N_id = 100,
  B = NULL,
  V = 3,
  groups = NULL,
  mode = "bernoulli",
  individual_predictors = NULL,
  dyadic_predictors = NULL,
  individual_effects = NULL,
  dyadic_effects = NULL
)
}
\arguments{
\item{N_id}{Number of individuals.}

\item{B}{List of matrices that hold intercept and offset terms. Log-odds. The first matrix should be  1 x 1 with the value being the intercept term.}

\item{V}{Number of blocking variables in B.}

\item{groups}{Dataframe of the block IDs of each individual for each variable in B.}

\item{mode}{Outcome mode: can be "bernoulli", "poisson", or "binomial".}

\item{individual_predictors}{An N_id by N_individual_parameters matrix of covariates.}

\item{dyadic_predictors}{An N_id by N_id by N_dyadic_parameters array of covariates.}

\item{individual_effects}{A 2 by N_individual_parameters matrix of slopes. The first row gives effects of focal characteristics (on out-degree). 
The second row gives effects of target characteristics (on in-degree).}

\item{dyadic_effects}{An N_dyadic_parameters vector of slopes.}
}
\value{
A list of objects including: network (an adjacency matrix of binary outcomes), tie_strength (an adjacency matrix with probability weights), 
group_ids (a vector of length N_id, giving the group of each individual), individual_predictors (the supplied covariate data is saved along with the network data), 
and dyadic_predictors (the supplied covariate data is saved along with the network data).
}
\description{
This is a function to simulate single layer network data with a stochastic block structure.
}
