% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/simulate_sbm_plus_srm_network_with_measurement_error.R
\name{simulate_sbm_plus_srm_network_with_measurement_bias}
\alias{simulate_sbm_plus_srm_network_with_measurement_bias}
\title{A function to simulate single layer directed networks using a combined stochastic block and social relations model}
\usage{
simulate_sbm_plus_srm_network_with_measurement_bias(
  N_id = 30,
  B = NULL,
  V = 3,
  groups = NULL,
  sr_mu = c(0, 0),
  sr_sigma = c(0.3, 1.5),
  sr_rho = 0.6,
  dr_mu = 0,
  dr_sigma = 1,
  dr_rho = 0.7,
  exposure_mu = 1.9,
  exposure_sigma = 0.01,
  exposure_max = 50,
  censoring_mu = 1.9,
  censoring_sigma = 0.01,
  N_trials = 20,
  outcome_mode = "binomial",
  link_mode = "logit",
  individual_predictors = NULL,
  dyadic_predictors = NULL,
  exposure_predictors = NULL,
  censoring_predictors = NULL,
  individual_effects = NULL,
  dyadic_effects = NULL,
  exposure_effects = NULL,
  censoring_effects = NULL
)
}
\arguments{
\item{N_id}{Number of individuals.}

\item{B}{List of matrices that hold intercept and offset terms. Log-odds. The first matrix should be  1 x 1 with the value being the intercept term.}

\item{V}{Number of blocking variables in B.}

\item{groups}{Dataframe of the block IDs of each individual for each variable in B.}

\item{sr_mu}{Mean vector for sender and receivier random effects. In most cases, this should be c(0,0).}

\item{sr_sigma}{A standard deviation vector for sender and receivier random effects. The first element controls node-level variation in out-degree, the second in in-degree.}

\item{sr_rho}{Correlation of sender-receiver effects (i.e., generalized reciprocity).}

\item{dr_mu}{Mean vector for dyadic random effects. In most cases, this should be c(0,0).}

\item{dr_sigma}{Standard deviation for dyadic random effects.}

\item{dr_rho}{Correlation of dyad effects (i.e., dyadic reciprocity).}

\item{exposure_mu}{Intercept term for log-odds of encounter.}

\item{exposure_sigma}{Standard deviation for exposure random effects.}

\item{exposure_max}{Max sample size of observations for a given focal.}

\item{censoring_mu}{Intercept term for log-odds of censoring.}

\item{censoring_sigma}{Standard deviation for censoring random effects.}

\item{N_trials}{Number of binomial trials in follow-up detectability experiment.}

\item{outcome_mode}{Outcome mode: only "binomial" is supported.}

\item{link_mode}{Link mode: can be "logit", "probit".}

\item{individual_predictors}{An N_id by N_individual_parameters matrix of covariates.}

\item{dyadic_predictors}{An N_id by N_id by N_dyadic_parameters array of covariates.}

\item{exposure_predictors}{An N_id by N_individual_parameters matrix of covariates.}

\item{censoring_predictors}{An N_id by N_individual_parameters matrix of covariates.}

\item{individual_effects}{A 2 by N_individual_parameters matrix of slopes. The first row gives effects of focal characteristics (on out-degree). 
The second row gives effects of target characteristics (on in-degree).}

\item{dyadic_effects}{An N_dyadic_parameters vector of slopes.}

\item{exposure_effects}{An N_parameters vector of slopes.}

\item{censoring_effects}{An N_parameters vector of slopes.}
}
\value{
A list of objects including: network (an adjacency matrix of binary outcomes), tie_strength (an adjacency matrix with probability weights), 
group_ids (a vector of length N_id, giving the group of each individual), individual_predictors (the supplied covariate data is saved along with the network data), 
and dyadic_predictors (the supplied covariate data is saved along with the network data).
}
\description{
This is a function to simulate single layer network data with a stochastic block structure, sender-receiver effects, and dyadic reciprocity. This function
is essentially the union of a social relations model and a stochastic block model.
}
