% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sr_match}
\alias{sr_match}
\title{Inexact Sequential Matching}
\usage{
sr_match(dat, d_vars, id_var, treat_var, pscores, trim = c(min = 0, max =
  1), mahalanobis = TRUE)
}
\description{
Inexact Sequential Matching
}
