% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sr_pscore}
\alias{sr_pscore}
\title{Imbens-Rubin Propensity Score}
\usage{
sr_pscore(dat, primary, secondary, id_var, min_ratios = c(lin = 2.71,
  quad = 2.71))
}
\description{
Imbens-Rubin Propensity Score
}
