% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRAND.R
\docType{package}
\name{STRAND}
\alias{STRAND}
\title{STRAND: Social network analysis and simulation in R using Stan}
\description{
The STRAND package provides three categories of important functions:
network simulation functions, network analysis functions, and summary functions
}
\section{Network simulation functions}{

The package can be used to simulate network data under stochastic block models, social relations models, and latent network models. 
These tools allow for simulation not only of realistic human social networks, but also allow researchers to simulate the effects of 
potential biases—like respondents falsely reporting ties or failing to recall real ties—on network level properties.
}

\section{Network analysis functions}{

Users can specify complex Bayesian social network models using simple lm() style syntax. 
Single-sampled self report data can be modeled using stochastic block models or the social relations model, with or with-out covariates. 
Double-sampled network data can be modeled using a latent network approach that accounts for inter-respondent disagreement.
}

\section{Summary functions}{

Users can summarize and visualize Stan results with a simple interface.
}

