% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_caterpillar_plot.R
\name{strand_caterpillar_plot}
\alias{strand_caterpillar_plot}
\title{A function to create a STRAND data object}
\usage{
strand_caterpillar_plot(
  results,
  submodels = NULL,
  normalized = FALSE,
  only_slopes = FALSE,
  only_technicals = FALSE,
  site = "BOB",
  export_as_table = FALSE
)
}
\arguments{
\item{results}{The results object from summarize_strand_results.}

\item{submodels}{Which submodels to plot? Supported: "False positive rate", "Recall of true ties", "Theta: question-order effects", "Focal effects: Out-degree", "Target effects: In-degree", "Dyadic effects", "Other estimates"}

\item{normalized}{Should effects be normalized? Do not use for correlations and variance terms.}

\item{only_slopes}{If TRUE then technical parameters are omited, and only slopes and intercepts are plotted.}

\item{only_technicals}{If TRUE then technical parameters are plotted, and slopes and intercepts are omited}

\item{site}{A string to indicate group or feildsite.}

\item{export_as_table}{If TRUE, the data are not plotted, and a data.frame is returned. This is useful for making complex ggplot objects with multiple groups/sites.}
}
\value{
A plot or dataframe
}
\description{
This function organizes network data and covariates into a form that can be used by STRAND for model fitting. All 
STRAND model fitting functions require their data to be supplied in the format exported here.
}
\examples{
\dontrun{
vis = strand_caterpillar_plot(res, submodel=c("Focal efffects: Out-degree",
                                              "Target effects: In-degree",
                                              "Dyadic effects","Other estimates"), 
                                   normalized=TRUE, 
                                   site="XY", 
                                   only_technicals=TRUE, 
                                   export_as_table=FALSE
                                    )
}

}
