% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_VPCs.R
\name{strand_VPCs}
\alias{strand_VPCs}
\title{Organize Stan output and provide summaries of variance parameters}
\usage{
strand_VPCs(
  input,
  n_partitions = 4,
  HPDI = 0.9,
  include_reciprocity = FALSE,
  mode = "cor",
  include_samples = FALSE,
  plot_pairs = FALSE,
  plot_type = "vpcs",
  plot_layer = NULL,
  verbose = FALSE
)
}
\arguments{
\item{input}{A STRAND model object with social relations model parameters.}

\item{n_partitions}{Should variance on latent scale be partioned into 3 factors (focal, target, dyadic+error) as in amen? Or 4 factors (focal, target, dyadic, error)? Or 5 factors (focal, target, dyadic reciprocal, dyadic noise, error)?}

\item{HPDI}{Highest Posterior Density Interval. Ranges in (0,1).}

\item{include_reciprocity}{Should reciprocity estimates be returned?.}

\item{mode}{Should the dyadic correlation "cor" be computed, or the dyadic covariance "cov", or the adjusted dyadic+error correlation "adj".}

\item{include_samples}{An indicator for the user to specify if raw samples, or only the summary statistics should be returned. Samples can take up a lot of space.}

\item{plot_pairs}{A Boolean to plot pairs of the variance terms.}

\item{plot_type}{Should the raw variances be plotted: "raw", or should the variance partition coefficients: "vpcs"?}

\item{plot_layer}{The name of the layer to plot.}

\item{verbose}{If TRUE prints the results.}
}
\value{
A STRAND results object.
}
\description{
This is a function to organize Stan output and provide summaries of key model parameters
}
\examples{
\dontrun{
res = strand_VPCs(input=fit)
}

}
