% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stats.R
\name{sum_stats}
\alias{sum_stats}
\title{A function to summarize samples}
\usage{
sum_stats(y, x, z)
}
\arguments{
\item{y}{A string.}

\item{x}{Samples.}

\item{z}{HPDI level}
}
\value{
A vector of summaries.
}
\description{
This is a simple helper function to summarize samples.
}
