% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_set_to_long.R
\name{target_set_to_long}
\alias{target_set_to_long}
\title{A function to map target or receiver predictors to a long-form dyad list.}
\usage{
target_set_to_long(X)
}
\arguments{
\item{X}{A data.frame.}
}
\value{
target_set_to_long(X)
}
\description{
This is an internal function.
}
