% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Callithrix_Data.R
\docType{data}
\name{Callithrix_Data}
\alias{Callithrix_Data}
\title{Data on Callithrix aggression}
\format{
An object of class \code{"list"}. This data object contains a subset of data from the original publication. The first sub-list contains data from the C-, or low food, condition.
The second sub-list contains data from the S++, or high food, condition. Both sub-lists contain data on dyadic aggression, normalized directed rank difference, opportunity to be observed, and sex.
}
\source{
\href{https://doi.org/10.5061/dryad.7pvmcvdtn}{Dryad}
}
\usage{
data(Callithrix_Data)
}
\description{
Foraging networks and social tolerance in a cooperatively breeding primate (Callithrix jacchus). See the reference to the original study below for full methodological information.
}
\examples{
data(Callithrix_Data)
}
\references{
De la Fuente, María Fernanda et al. (2021), Foraging networks and social tolerance in a cooperatively breeding primate (Callithrix jacchus), Dryad, Dataset.
}
\keyword{datasets}
