% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_strand_venv.R
\name{create_strand_venv}
\alias{create_strand_venv}
\title{A function to create a Python venv for STRAND to use for NumPyro models}
\usage{
create_strand_venv(rebuild = FALSE, verbose = TRUE, mcmc_parameters)
}
\arguments{
\item{rebuild}{The user needs to run this once as TRUE to rebuild. Inside of STRAND we will load the pre-built venv.}

\item{verbose}{If TRUE, it will print details of GPU detections.}

\item{mcmc_parameters}{A parameter list to control NUTS sampling. NumPyro uses the same control set as Stan.}
}
\description{
This is a simple helper function to get Python running smoothly
}
