% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_inputs}
\alias{sample_inputs}
\title{Sample security inputs for examples and testing}
\format{
A data frame with 31500 rows and 9 variables:
\describe{
  \item{id}{security identifier}
  \item{date}{input date}
  \item{average_volume}{measurement of average security trading volume, in shares}
  \item{alpha_1}{sample numeric alpha input}
  \item{alpha_2}{sample numeric alpha input}
  \item{factor_1}{sample numeric factor input}
  \item{factor_2}{sample numeric factor input}
  \item{factor_3}{sample numeric factor input}
  \item{factor_4}{sample numeric factor input}
}
}
\usage{
sample_inputs
}
\description{
A dataset containing sample (fake) security input data for 500 securities and
63 weekdays, from 2019-01-02 to 2019-03-29.
}
\keyword{datasets}
