% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_pricing}
\alias{sample_pricing}
\title{Sample security pricing data for examples and testing}
\format{
A data frame with 31980 rows and 8 variables:
\describe{
  \item{date}{Pricing date.}
  \item{id}{Security identifier.}
  \item{price_unadj}{The unadjusted price of the security.}
  \item{prior_close_unadj}{The unadjusted prior closing price of the security.}
  \item{dividend_unadj}{The dividend for the security on an unadjusted basis, if any.}
  \item{distribution_unadj}{The distribution (e.g., spin-off) for the
  security on an unadjusted basis (note that there is no spin-off information
  in this dataset, so all values are zero).}
  \item{volume}{Trading volume for the security, in shares.}
  \item{adjustment_ratio}{The adjustment ratio for the security. For example,
  AAPL has an adjustment ratio of 0.25 to account for its 4:1 split on
  2020-08-31.}
}
}
\usage{
data(sample_pricing)
}
\description{
A dataset containing sample security pricing data for 492 securities and 65
weekdays, from 2020-06-01 to 2020-08-31. This data was downloaded using the
\href{https://api.tiingo.com/}{Tiingo Stock API} and is redistributed with
permission.
}
\details{
Full code for reconstructing the dataset can be found in the
\href{https://github.com/strand-tech/pystrand}{pystrand} repository.
}
\keyword{datasets}
