% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_secref}
\alias{sample_secref}
\title{Sample security reference data for examples and testing}
\format{
A data frame with 500 rows and 4 variables:
\describe{
  \item{id}{security identifier}
  \item{symbol}{human-readable trading symbol}
  \item{category_1}{categorical variable with values A-F}
  \item{category_2}{categorical variable with values A-L}
}
}
\usage{
sample_secref
}
\description{
A dataset containing sample (fake) security reference data for 500 securities.
}
\keyword{datasets}
