% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sbm_network.R
\name{simulate_sbm_network}
\alias{simulate_sbm_network}
\title{A function to simulate single layer directed networks with a stochastic block structure}
\usage{
simulate_sbm_network(
  N_id = 100,
  B = NULL,
  V = 3,
  groups = NULL,
  error_sigma = 0.01,
  outcome_mode = "bernoulli",
  link_mode = "logit",
  individual_predictors = NULL,
  dyadic_predictors = NULL,
  individual_effects = NULL,
  dyadic_effects = NULL
)
}
\arguments{
\item{N_id}{Number of individuals.}

\item{B}{List of matrices that hold intercept and offset terms. Log-odds. The first matrix should be  1 x 1 with the value being the intercept term.}

\item{V}{Number of blocking variables in B.}

\item{groups}{Dataframe of the block IDs of each individual for each variable in B.}

\item{error_sigma}{Standard deviation for measurement error in Gaussian models.}

\item{outcome_mode}{Outcome mode: can be "bernoulli", "poisson", "binomial", or "gaussian".}

\item{link_mode}{Link mode: can be "logit", "probit", "log", or "identity". For poisson, you must use log. For Gaussian you must use identity.}

\item{individual_predictors}{An N_id by N_individual_parameters matrix of covariates.}

\item{dyadic_predictors}{An N_id by N_id by N_dyadic_parameters array of covariates.}

\item{individual_effects}{A 2 by N_individual_parameters matrix of slopes. The first row gives effects of focal characteristics (on out-degree). 
The second row gives effects of target characteristics (on in-degree).}

\item{dyadic_effects}{An N_dyadic_parameters vector of slopes.}
}
\value{
A list of objects including: network (an adjacency matrix of binary outcomes), tie_strength (an adjacency matrix with probability weights), 
group_ids (a vector of length N_id, giving the group of each individual), individual_predictors (the supplied covariate data is saved along with the network data), 
and dyadic_predictors (the supplied covariate data is saved along with the network data).
}
\description{
This is a function to simulate single layer network data with a stochastic block structure.
}
\examples{
\dontrun{
library(igraph)
V = 1            # One blocking variable
G = 3            # Three categories in this variable
N_id = 100       # Number of people

clique = sample(1:3, N_id, replace=TRUE)
B = matrix(-6, nrow=G, ncol=G)
diag(B) = -1.5 # Block matrix

B[1,3] = -5
B[3,2] = -4

A = simulate_sbm_network(N_id = N_id, B=list(B=B), V=V, groups = data.frame(clique=factor(clique)),
                         individual_predictor=matrix(rnorm(N_id, 0, 1), nrow=N_id, ncol=1), 
                         individual_effects=matrix(c(1.7, 0.3),ncol=1, nrow=2),
                         outcome_mode="bernoulli"
                               )

Net = graph_from_adjacency_matrix(A$network, mode = c("directed"))
V(Net)$color = c("turquoise4","gray13", "goldenrod3")[A$group_ids$clique]

plot(Net, edge.arrow.size =0.1, edge.curved = 0.3, vertex.label=NA, vertex.size = 5)
}

}
