% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\docType{package}
\name{strand-package}
\alias{strand-package}
\alias{strand}
\title{strand: a framework for investment strategy simulation}
\description{
The strand package provides a framework for performing discrete (share-level)
simulations of investment strategies. Simulated portfolios optimize exposure
to an input signal subject to constraints such as position size and factor
exposure.

For an introduction to running simulations using the package, see
\code{vignette("strand")} and the documentation for the \code{\link{Simulation}} class.
}
\author{
Jeff Enos \email{jeff@strand.tech} and David Kane \email{david@strand.tech}
}
