% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\docType{package}
\name{strand-package}
\alias{strand-package}
\alias{strand}
\title{strand: a framework for investment strategy simulation}
\description{
For an introduction to strand see the package vignette:
\code{vignette("strand", package = "strand")}
}
\author{
Jeff Enos \email{jeff@strand.tech} and David Kane \email{david@strand.tech}
}
