% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_bm_results.R
\name{summarize_bm_results}
\alias{summarize_bm_results}
\title{Organize Stan output and provide summaries of model parameters}
\usage{
summarize_bm_results(input, include_samples = TRUE)
}
\arguments{
\item{input}{A STRAND model object, obtained by fitting a block model.}

\item{include_samples}{Should raw samples be returned? Or only the summary statistics? Samples can take up a lot of space.}
}
\value{
A STRAND results object including summary table, a summary list, and samples.
}
\description{
This is a function to organize Stan output and provide summaries of key model parameters
}
\examples{
\dontrun{
res = summarize_bm_results(input=fit)
}

}
